+++ CHANGES.rst
@@ -91,6 +91,9 @@ Fixes
   to the shell.
   (`#1597 <https://github.com/git-cola/git-cola/pull/1597>`_)
 
+* Support for ``$LANGUAGE``, ``$LC_ALL``, ``$LC_MESSAGES`` and ``$LANG`` was improved.
+  (`#519972 <https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=519972>`_)
+
 Development
 -----------
 * ``isort`` is now used for code formatting.